+++ src/main/java/com/buuz135/replication/client/MatterTooltipClientComponent.java
@@ -74,7 +74,7 @@ public class MatterTooltipClientComponent implements ClientTooltipComponent {
         guiGraphics.pose().pushPose();
         float scale = 0.5f;
         guiGraphics.pose().scale(scale, scale, scale);
-        var number = NumberUtils.getFormatedBigNumber(Mth.ceil(instance.getAmount()));
+        var number = NumberUtils.getFormatedBigNumber(instance.getAmount());
         guiGraphics.drawCenteredString(Minecraft.getInstance().font, number, (int) ((x + (glitch ? 1 : 0) + 8) / scale), (int) ((y + (glitch ? 1 : 0) + 15) / scale), new Color(color[0], color[1], color[2], color[3]).getRGB());
         guiGraphics.pose().popPose();
         /*OLD ENCODED NUMBERS guiGraphics.blit(BAR, x + (glitch ? 1 : 0), y +(glitch ? 1 : 0), 0,56, 3, 15, 128, 128);